+++ golang/internal/config/config.go
@@ -32,7 +32,7 @@ type CommonConfiguration struct {
 	Name                 string        `yaml:"name"                  env:"NAME"                    env-default:"dagent-go"`
 	Debug                bool          `yaml:"debug"                 env:"DEBUG"                   env-default:"false"`
 	ImportContainerImage string        `yaml:"importContainerImage"  env:"IMPORT_CONTAINER_IMAGE"  env-default:"rclone/rclone:1.57.0"`
-	IngressRootDomain    string        `yaml:"ingressRootDomain"     env:"INGRESS_ROOT_DOMAIN"     env-default:""`
+	RootDomain           string        `yaml:"rootDomain"            env:"ROOT_DOMAIN"             env-default:""`
 	ReadHeaderTimeout    time.Duration `yaml:"readHeaderTimeout"     env:"READ_HEADER_TIMEOUT"     env-default:"15s"`
 	// DefaultRegistry container registry used for container name expansion
 	DefaultRegistry string `yaml:"registry"             env:"DEFAULT_REGISTRY"                 env-default:"index.docker.io"`