+++ CHANGELOG.md
@@ -1,6 +1,31 @@
 # CHANGELOG
 
 
+<a name="0.6.1"></a>
+## [0.6.1](https://github.com/dyrector-io/dyrectorio/compare/0.6.0...0.6.1) (2023-07-14)
+
+### Fix
+
+* failing deployment token migration ([#750](https://github.com/dyrector-io/dyrectorio/issues/750))
+* none container logger & port validation & moved in use warnings ([#741](https://github.com/dyrector-io/dyrectorio/issues/741))
+* **(ci):** fix the unrecognized error arround external deploy workflow ([#744](https://github.com/dyrector-io/dyrectorio/issues/744))
+* **(crux-ui):** add view buttons to deployment list rows ([#743](https://github.com/dyrector-io/dyrectorio/issues/743))
+* **(crux-ui):** change confirm button color ([#747](https://github.com/dyrector-io/dyrectorio/issues/747))
+
+
+<a name="0.6.0"></a>
+## [0.6.0](https://github.com/dyrector-io/dyrectorio/compare/0.5.5...0.6.0) (2023-07-12)
+
+### Feat
+
+* unchecked registry use local images ([#745](https://github.com/dyrector-io/dyrectorio/issues/745))
+
+### Fix
+
+* remove n refactor todos ([#742](https://github.com/dyrector-io/dyrectorio/issues/742))
+* deployment token name validation ([#740](https://github.com/dyrector-io/dyrectorio/issues/740))
+
+
 <a name="0.5.5"></a>
 ## [0.5.5](https://github.com/dyrector-io/dyrectorio/compare/0.5.4...0.5.5) (2023-07-10)